extract to_isolang
