+++ portality/lib/coerce.py
@@ -23,42 +23,6 @@ def date_str(in_format=None, out_format=None):
     return datify
 
 
-def to_isolang(output_format=None):
-    """
-    :param output_format: format from input source to putput.  Must be one of:
-        * alpha3
-        * alt3
-        * alpha2
-        * name
-        * fr
-    Can be a list in order of preference, too
-    ~~-> Languages:Data~~
-    :return:
-    """
-    # delayed import, since we may not always want to load the whole dataset for a dataobj
-    from portality.lib import isolang as dataset
-
-    # sort out the output format list
-    if output_format is None:
-        output_format = ["alpha3"]
-    if not isinstance(output_format, list):
-        output_format = [output_format]
-
-    def isolang(val):
-        if val is None:
-            return None
-        l = dataset.find(val)
-        if l is None:
-            raise ValueError("Unable to find iso code for language {x}".format(x=val))
-        for f in output_format:
-            v = l.get(f)
-            if v is None or v == "":
-                continue
-            return v.upper()
-
-    return isolang
-
-
 def to_currency_code(val):
     """
     ~~-> Currencies:Data~~
